GCW-1910 Singleton Item dispatch to GC order
